requirements.txt
# Face Recognition System - Exact Version Requirements
# Tested with Python 3.12.7

# Core face detection and recognition
ultralytics==8.3.245
huggingface-hub==0.36.0
supervision==0.27.0
face-recognition==1.3.0
retina-face==0.0.17
deepface==0.0.96

# Image processing
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.11.0.86

# Deep learning frameworks
torch==2.9.1
torchvision==0.24.1
tf-keras>=2.20.0  # Required for retina-face with TensorFlow 2.20+

# Gemini Live API (Client-to-Server)
websockets>=12.0
pyaudio>=0.2.14
google-genai>=0.2.0

# Google Sheets integration for Smart Attendance
gspread>=5.12.0
oauth2client>=4.1.3
google-auth>=2.45.0  # Required for compatibility with google-genai

